bot_service/requirements.txt
python-telegram-bot==20.3
redis==5.2.0
httpx==0.24.1
anyio==4.11.0
